perf: move network test logic to common
